app/algorithms/requirements.txt
firebase_admin
certifi==2020.12.5
chardet==4.0.0
Cython==0.29.22
dataclasses==0.6
hdbscan==0.8.27
ibm-cloud-sdk-core==3.5.2
ibm-watson==5.1.0
idna==2.10
joblib==1.0.1
nose==1.3.7
numpy==1.19.5
PyJWT==2.0.1
python-dateutil==2.8.1
PyYAML==5.4.1
Random-Word==1.0.7
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
torch==1.7.1
torchtext==0.8.1
tqdm==4.57.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.48.0
